requirements.txt
pillow
pystray
readmdict
levenshtein
tkinterweb